Move out Python/R related configurations to separate objects
