+++ core/src/main/scala/org/apache/spark/api/r/RRunner.scala
@@ -27,7 +27,7 @@ import scala.util.Try
 import org.apache.spark._
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.internal.Logging
-import org.apache.spark.internal.config._
+import org.apache.spark.internal.config.R._
 import org.apache.spark.util.Utils
 
 /**
