+++ core/src/main/scala/org/apache/spark/deploy/RRunner.scala
@@ -26,7 +26,7 @@ import org.apache.hadoop.fs.Path
 
 import org.apache.spark.{SparkException, SparkUserAppException}
 import org.apache.spark.api.r.{RBackend, RUtils, SparkRDefaults}
-import org.apache.spark.internal.config._
+import org.apache.spark.internal.config.R._
 import org.apache.spark.util.RedirectThread
 
 /**
